패키지 구조 변경 및 유효성 검사(형식)
api 통신 없이 자체 검사
