+++ app/src/main/AndroidManifest.xml
@@ -25,12 +25,12 @@
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
         <activity
-            android:name=".features.intro.IntroActivity"
+            android:name=".features.intro.presentation.IntroActivity"
             android:exported="true"
             android:theme="@style/FullScreenTheme"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".features.intro.onboarding.presentation.OnBoardingActivity"
+            android:name="com.example.king_bob_nae.features.onboarding.presentation.OnBoardingActivity"
             android:exported="true"
             android:theme="@style/FullScreenTheme">
             <intent-filter>
