--- app/src/main/java/com/example/king_bob_nae/features/intro/IntroActivity.kt
@@ -1,7 +1,0 @@
-package com.example.king_bob_nae.features.intro
-
-import com.example.king_bob_nae.R
-import com.example.king_bob_nae.base.BaseActivity
-import com.example.king_bob_nae.databinding.ActivitySigninBinding
-
-class IntroActivity : BaseActivity<ActivitySigninBinding>(R.layout.activity_signin)
