--- app/src/main/java/com/example/king_bob_nae/features/intro/onboarding/presentation/OnBoardingFirstFragment.kt
@@ -1,19 +1,0 @@
-package com.example.king_bob_nae.features.intro.onboarding.presentation
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.king_bob_nae.R
-import com.example.king_bob_nae.base.BaseFragment
-import com.example.king_bob_nae.databinding.FragmentOnboardingFirstBinding
-
-class OnBoardingFirstFragment : BaseFragment<FragmentOnboardingFirstBinding>(R.layout.fragment_onboarding_first) {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-}
