+++ app/src/main/java/com/example/king_bob_nae/features/onboarding/presentation/OnBoardingLastFragment.kt
@@ -1,0 +1,19 @@
+package com.example.king_bob_nae.features.onboarding.presentation
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.king_bob_nae.R
+import com.example.king_bob_nae.base.BaseFragment
+import com.example.king_bob_nae.databinding.FragmentOnboardingLastBinding
+
+class OnBoardingLastFragment : BaseFragment<FragmentOnboardingLastBinding>(R.layout.fragment_onboarding_last) {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return super.onCreateView(inflater, container, savedInstanceState)
+    }
+}
