--- app/src/main/java/com/example/king_bob_nae/features/signup/nickname/presentation/SignUpNicknameFragment.kt
@@ -1,32 +1,0 @@
-package com.example.king_bob_nae.features.signup.nickname.presentation
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.navigation.findNavController
-import com.example.king_bob_nae.R
-import com.example.king_bob_nae.base.BaseFragment
-import com.example.king_bob_nae.databinding.FragmentSignUpNicknameBinding
-
-class SignUpNicknameFragment :
-    BaseFragment<FragmentSignUpNicknameBinding>(R.layout.fragment_sign_up_nickname) {
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initView()
-    }
-
-    private fun initView() {
-        binding.apply {
-            btnSignUpNickBack.setOnClickListener {
-                it.findNavController()
-                    .navigate(R.id.action_signUpNicknameFragment_to_signUpPasswdFragment)
-            }
-            btnSignUpNickNext.setOnClickListener {
-                it.findNavController()
-                    .navigate(R.id.action_signUpNicknameFragment_to_signUpLoadingFragment)
-            }
-        }
-    }
-}
