+++ app/src/main/res/navigation/intro_navigation_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/introFragment"
-        android:name="com.example.king_bob_nae.features.intro.IntroFragment"
+        android:name="com.example.king_bob_nae.features.intro.presentation.IntroFragment"
         android:label="IntroFragment">
         <action
             android:id="@+id/action_loginFragment_to_signInFragment"
@@ -28,7 +28,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpEmailFragment2"
-        android:name="com.example.king_bob_nae.features.signup.email.presentation.SignUpEmailFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.email.presentation.SignUpEmailFragment"
         android:label="SignUpEmailFragment" >
         <action
             android:id="@+id/action_signUpEmailFragment2_to_introFragment"
@@ -39,7 +39,7 @@
     </fragment>
     <fragment
         android:id="@+id/checkEmailFragment2"
-        android:name="com.example.king_bob_nae.features.reset.check_email.presentation.CheckEmailFragment"
+        android:name="com.example.king_bob_nae.features.intro.reset.check_email.presentation.CheckEmailFragment"
         android:label="CheckEmailFragment" >
         <action
             android:id="@+id/action_checkEmailFragment2_to_signInFragment"
@@ -50,7 +50,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpCheckCertificationFragment"
-        android:name="com.example.king_bob_nae.features.signup.certification.presentation.SignUpCheckCertificationFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.certification.presentation.SignUpCheckCertificationFragment"
         android:label="SignUpCheckCertificationFragment" >
         <action
             android:id="@+id/action_signUpCheckCertificationFragment_to_signUpEmailFragment2"
@@ -61,7 +61,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpPasswdFragment"
-        android:name="com.example.king_bob_nae.features.signup.password.presentation.SignUpPasswdFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.password.presentation.SignUpPasswdFragment"
         android:label="SignUpPasswdFragment" >
         <action
             android:id="@+id/action_signUpPasswdFragment_to_signUpEmailFragment2"
@@ -72,7 +72,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpNicknameFragment"
-        android:name="com.example.king_bob_nae.features.signup.nickname.presentation.SignUpNicknameFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.nickname.presentation.SignUpNicknameFragment"
         android:label="SignUpNicknameFragment" >
         <action
             android:id="@+id/action_signUpNicknameFragment_to_signUpPasswdFragment"
@@ -83,7 +83,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpLoadingFragment"
-        android:name="com.example.king_bob_nae.features.signup.loading.presentation.SignUpLoadingFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.loading.presentation.SignUpLoadingFragment"
         android:label="SignUpLoadingFragment" >
         <action
             android:id="@+id/action_signUpLoadingFragment_to_signUpSelectCharacterFragment"
@@ -91,7 +91,7 @@
     </fragment>
     <fragment
         android:id="@+id/signUpSelectCharacterFragment"
-        android:name="com.example.king_bob_nae.features.signup.select_character.presentation.SignUpSelectCharacterFragment"
+        android:name="com.example.king_bob_nae.features.intro.signup.select_character.presentation.SignUpSelectCharacterFragment"
         android:label="SignUpSelectCharacterFragment" >
         <action
             android:id="@+id/action_signUpSelectCharacterFragment_to_signInFragment"
@@ -99,7 +99,7 @@
     </fragment>
     <fragment
         android:id="@+id/checkCertificationFragment"
-        android:name="com.example.king_bob_nae.features.reset.check_certification.presentation.CheckCertificationFragment"
+        android:name="com.example.king_bob_nae.features.intro.reset.check_certification.presentation.CheckCertificationFragment"
         android:label="CheckCertificationFragment" >
         <action
             android:id="@+id/action_checkCertificationFragment_to_checkEmailFragment2"
@@ -110,7 +110,7 @@
     </fragment>
     <fragment
         android:id="@+id/resetPasswdFragment"
-        android:name="com.example.king_bob_nae.features.reset.reset_passwd.presentation.ResetPasswdFragment"
+        android:name="com.example.king_bob_nae.features.intro.reset.reset_passwd.presentation.ResetPasswdFragment"
         android:label="ResetPasswdFragment" >
         <action
             android:id="@+id/action_resetPasswdFragment_to_checkEmailFragment22"
